Create kappen group
The kappen group is a calicut based architecture and interior design firm concentrating on projects leading from individual homes to commercial buildings. Our company delivers clients with top quality services in Architecture. The main aim of our firm is to practice architecture as a relevant and responsive medium that can accommodate a theoretically significant design while creating the latest architectural design and theory. Our architects not only run behind the traditional designs but always come up with modern design concepts. As a famous <a href=The kappen group is a calicut based architecture and interior design firm concentrating on projects leading from individual homes to commercial buildings. Our company delivers clients with top quality services in Architecture. The main aim of our firm is to practice architecture as a relevant and responsive medium that can accommodate a theoretically significant design while creating the latest architectural design and theory. Our architects not only run behind the traditional designs but always come up with modern design concepts. As a famous architect Calicut , we work in several industries to design and build the architectures. It may be the residential, industrial, commercial, cultural, medical, and educational institutions.

>architect Calicut</a>, we work in several industries to design and build the architectures. It may be the residential, industrial, commercial, cultural, medical, and educational institutions.